backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.10.0",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "cryptography==46.0.3",
    "defusedxml==0.7.1",
    "django==5.2.8",
    "django-filter==25.2",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "djoser==2.3.3",
    "idna==3.11",
    "oauthlib==3.3.1",
    "pillow==12.0.0",
    "pycparser==2.23",
    "pyjwt==2.10.1",
    "python-dotenv==1.2.1",
    "python3-openid==3.2.0",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "social-auth-app-django==5.6.0",
    "social-auth-core==4.8.1",
    "sqlparse==0.5.3",
    "urllib3==2.5.0",
]
[tool.ruff]
line-length = 79
target-version = "py311"
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    ".venv",
    "venv",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
]
ignore = [
    "E501",
    "B008",
]

[tool.ruff.lint.isort]
known-first-party = ["api", "recipes", "users", "foodgram"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
reportIncompatibleVariableOverride = false
reportIncompatibleMethodOverride = false
reportAttributeAccessIssue = false
